fix: txpool size change to update normal and foundrows pool
