+++ go/vt/vttablet/endtoend/call_test.go
@@ -20,7 +20,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-
 	"github.com/stretchr/testify/require"
 
 	"vitess.io/vitess/go/vt/vttablet/endtoend/framework"
@@ -136,6 +135,7 @@ func TestCallProcedureLeakTx(t *testing.T) {
 
 func TestCallProcedureChangedTx(t *testing.T) {
 	client := framework.NewClient()
+	defer client.Release()
 
 	_, err := client.Execute(`call proc_tx_begin()`, nil)
 	require.EqualError(t, err, "Transaction not concluded inside the stored procedure, leaking transaction from stored procedure is not allowed (CallerID: dev)")
