+++ go/vt/vttablet/endtoend/config_test.go
@@ -23,7 +23,6 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-
 	"github.com/stretchr/testify/require"
 
 	"vitess.io/vitess/go/sqltypes"
@@ -71,6 +70,48 @@ func TestStreamPoolSize(t *testing.T) {
 
 	vstart := framework.DebugVars()
 	verifyIntValue(t, vstart, "StreamConnPoolCapacity", 1)
+}
+
+// TestTxPoolSize starts 2 transactions, one in normal pool and one in found rows pool of transaction pool.
+// Changing the pool size to 1, we verify that the pool size is updated and the pool is full when we try to acquire next transaction.
+func TestTxPoolSize(t *testing.T) {
+	vstart := framework.DebugVars()
+
+	verifyIntValue(t, vstart, "TransactionPoolCapacity", 20)
+	verifyIntValue(t, vstart, "FoundRowsPoolCapacity", 20)
+
+	client1 := framework.NewClient()
+	err := client1.Begin( /* found rows pool*/ false)
+	require.NoError(t, err)
+	defer client1.Rollback()
+	verifyIntValue(t, framework.DebugVars(), "TransactionPoolAvailable", framework.FetchInt(vstart, "TransactionPoolAvailable")-1)
+
+	client2 := framework.NewClient()
+	err = client2.Begin( /* found rows pool*/ true)
+	require.NoError(t, err)
+	defer client2.Rollback()
+	verifyIntValue(t, framework.DebugVars(), "FoundRowsPoolAvailable", framework.FetchInt(vstart, "FoundRowsPoolAvailable")-1)
+
+	revert := changeVar(t, "TxPoolSize", "1")
+	defer revert()
+	vend := framework.DebugVars()
+	verifyIntValue(t, vend, "TransactionPoolAvailable", 0)
+	verifyIntValue(t, vend, "TransactionPoolCapacity", 1)
+	verifyIntValue(t, vend, "FoundRowsPoolAvailable", 0)
+	verifyIntValue(t, vend, "FoundRowsPoolCapacity", 1)
+	assert.Equal(t, 1, framework.Server.TxPoolSize())
+
+	client3 := framework.NewClient()
+
+	// tx pool - normal
+	err = client3.Begin( /* found rows pool*/ false)
+	require.ErrorContains(t, err, "connection limit exceeded")
+	compareIntDiff(t, framework.DebugVars(), "Errors/RESOURCE_EXHAUSTED", vstart, 1)
+
+	// tx pool - found rows
+	err = client3.Begin( /* found rows pool*/ true)
+	require.ErrorContains(t, err, "connection limit exceeded")
+	compareIntDiff(t, framework.DebugVars(), "Errors/RESOURCE_EXHAUSTED", vstart, 2)
 }
 
 func TestDisableConsolidator(t *testing.T) {
